Call validateFormWithLowPriority with correct values (#2025)
- When using `setValues` `dispatch` is called with type: SET_VALUES and
payload: values. However the state update happens asynchronously so when
`validateFormWithLowPriority` is called, `state.values` still hasn't
been updated.
- This fix improves the assertion for the `setValues`
validation to test to confirm that the expected value is given to
Formik's validate function.
- The fix itself is to call `validateFormWithLowPriority` with the
`values` parameter given to `setValues` itself, instead of waiting for
the `useReducer` to update the state. Currently it does not seem to be
possible to wait for `dispatch` to have completed.